[dev.regabi] cmd/compile: avoid OCONVNOP lvalues
It's a bit surprising that you can have an OCONVNOP as an lvalue.
Make that stop.

Passes buildall w/ toolstash -cmp.
That's not guaranteed - test/closure2.go compiles a little differently,
because the side effects are slightly reordered relative to the non-side-effects.
But weird cases like that don't happen in std cmd.

Change-Id: Ibd88440888137f5379f2f6931b47310ef6d14627