[dev.typeparams] cmd/compile/internal/types2: adjustments toward matching compiler error messages
In order to get types2 usable by the compiler, we need to pass all the compiler
tests with respect to error messages. Sometimes the compiler error messages are
better, sometimes the types2 error messages are better. Where we can, we decide
on a case-by-case basis; but sometimes, for expediency's sake, we just choose
the compiler error message as it may reduce the amount of tests that we need
to update.

This CL introduces a temporary new Config flag: CompilerErrorMessages.
If set, the typechecker emits an error message that matches the expected
errors in the tests most easily. Eventually, we need to get rid of this
flag by either adjusting the typechecker errors or the test cases; t.b.d.
on a case-by-case basis.

This CL also adjust a few error typechecker error messages already.

Change-Id: I9d4e491efadf87e999fc0d5b5151ec02a059f891