cmd/go: add documentation for the -overlay flag
Also add -overlay to the Go 1.16 release notes.

(Question for my reviewers: the -overlay flag also applies to the
'go mod' command, but I can't seem to find a good place to put
the documentation for the flag because each of the go mod subcommands
have their own documentation... any suggestions for where that should
go?

Fixes #39958

Change-Id: Ifd397549e368b255e7b8800986cfa0563a942af5