[dev.regabi] cmd/compile: add ConstExpr
Currently, we represent constant-folded expressions with Name, which
is suboptimal because Name has a lot of fields to support declared
names (which are irrelevant to constant-folded expressions), while
constant expressions are very common.

This CL introduces a new lightweight ConstExpr type that simply wraps
an existing expression and associates it with a value.

Change-Id: If5781be346351c30b2228807211b5e57f777c506