[dev.regabi] cmd/compile: add RefersTo and DeclaredBy helpers
Currently, because we use the same *Name to represent both declaration
and use of an object, there's a lot of code that uses "n1 == n2" to
check if an expression is a use of a particular variable. If we want
to introduce a separate IR node to represent object use, we need to
update this code.

The simplest way would be to write "n1.Name() == n2". However, this
has two issues: (1) it's easy to accidentally write "n1 == n2.Name()"
(which will typecheck, but do the wrong thing), and (2) I intend to
make the default Name() implementation panic so it can *only* be used
for nodes associated with declared objects.

So this CL instead introduces RefersTo, so users can write
"ir.RefersTo(n1, n2)", which will be less prone to misuse. It also
rewrites a few AST walks to use this function, along with refactoring
them to use ir.Inspect.

Finally, this CL also introduces DeclaredBy, which tests whether an
expression refers to a variable declared by a particular statement,
another common operation. The next CL rewrites code to use it.

Passes buildall w/ toolstash -cmp.

Updates #42981.
Updates #42990.

Change-Id: Ia816c124446e9067645d5820a8163f295968794f