[dev.fuzz] internal/fuzz: read and write interesting values to GOFUZZCACHE
GOFUZZCACHE is a new environment variable pointing to a directory
where 'go test -fuzz' may read and write "interesting" values for fuzz
testing.

GOFUZZCACHE may be read and written with 'go env'. The directory may
be removed with 'go clean -fuzzcache'.

Files in the cache are named $GOFUZZCACHE/$pkg/$test/$hash where $pkg
is the package path containing the test, $test is the test target
name, and $hash is the SHA-256 hash of the data in the file.

We don't yet have a heuristic for determining whether values are
"interesting". For now, each worker will run for a short time, then
claim that its last mutated value is interesting.

Change-Id: I2032cf8e6c92f715cf36a9fc6a550acf666d2382