runtime: limit maximum netbsd kevent delay
Issue 42515 tracks a still-unknown netbsd bug that prevents netpoll from
waking with netpollBreak, which could result in missing timers for an
unbounded amount of time, as netpoll can't restart with a shorter delay
when an earlier timer is added.

Prior to CL 232298, sysmon could detect these overrun timers and manually
start an M to run them. With this fallback gone, the bug actually
prevents timer execution indefinitely.

As a workaround, we bound netpoll delay to a maximum delay of 10ms. This
is equivalent to the maximum standard sysmon sleep and thus equivalent
in the max overrun it allows to prior to CL 232298.

The huge downside to the workaround is that applications with timers can
no longer go into deep sleep. Previously, sysmon could deep sleep if
everything was idle. That is not the case here.

Updates #42515

Change-Id: I8391f5b9dabef03dd1d94c50b3b4b3bd4f889e66