[dev.regabi] cmd/compile: cleanup for splitting package gc
This CL makes a few adjustments to the API between different
logical parts of package gc, so that future automated CLs can
split package gc into a few different packages.

This CL is focused primarily on cleanly separating the code
that will end up in package typecheck from the code that
will not.

Change-Id: I652068776205491416827d156cb4cc1d0df74f54